requirements/base.txt
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clr-loader==0.2.6
colorama==0.4.6
contextlib2==21.6.0
coverage==7.5.2
cryptography==42.0.7
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.0.6
django-extensions==3.2.3
django-ninja==1.1.0
django-ninja-extra==0.20.3
django-ninja-jwt==5.2.10
django-redis==5.4.0
dnspython==2.6.1
email_validator==2.1.1
Faker==22.0.0
fonttools==4.53.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gunicorn==21.2.0
httmock==1.4.0
idna==3.7
iniconfig==2.0.0
injector==0.21.0
packaging==24.0
phonenumbers==8.13.37
pillow==10.3.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pytest==8.2.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pythonnet==3.0.3
redis==4.3.4
requests==2.31.0
rsa==4.9
ruff==0.1.8
simplejson==3.16.0
six==1.12.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
wrapt==1.16.0
